+++ crates/messaging/affinidi-messaging-sdk/src/messages/list.rs
@@ -1,5 +1,10 @@
 use super::{Folder, MessageList};
-use crate::{ATM, errors::ATMError, messages::SuccessResponse, profiles::ATMProfile};
+use crate::{
+    ATM,
+    errors::{ATMError, HttpStatusError},
+    messages::SuccessResponse,
+    profiles::ATMProfile,
+};
 use sha256::digest;
 use std::sync::Arc;
 use tracing::{Instrument, Level, debug, span};
@@ -55,19 +60,8 @@ impl ATM {
                     ATMError::TransportError(format!("Could not send list_messages request: {e:?}"))
                 })?;
 
-            let status = res.status();
-            debug!("API response: status({})", status);
-
-            let body = res
-                .text()
-                .await
-                .map_err(|e| ATMError::TransportError(format!("Couldn't get body: {e:?}")))?;
-
-            if !status.is_success() {
-                return Err(ATMError::TransportError(format!(
-                    "Status not successful. status({status}), response({body})"
-                )));
-            }
+            debug!("API response: status({})", res.status());
+            let body = HttpStatusError::check_response("list messages", res).await?;
 
             let body =
                 serde_json::from_str::<SuccessResponse<MessageList>>(&body).map_err(|e| {